Split src/server/index.js
index.js is the entrypont for the server.
client.js handles serving the React.js client build.
config.js contains server configuration values.
The api directory contains our REST API endpoints.
