+++ src/server/api/v1/v1.js
@@ -1,0 +1,11 @@
+const express = require("express");
+const router = express.Router();
+
+const exampleEndpoints = require("./exampleEndpoints.js");
+
+router.use(exampleEndpoints);
+router.use((req, res) => {
+	res.sendStatus(404);
+});
+
+module.exports = router;
