+++ src/server/config.js
@@ -1,0 +1,6 @@
+const path = require("path");
+
+const PORT = process.env.PORT || 8080;
+const SITE_ROOT = process.env.SITE_ROOT || path.join(__dirname, "../../dist");
+
+module.exports = { PORT, SITE_ROOT };
